io: Move `io::der` unit tests to integration tests.
These are testing an API that is exposed publicly, so make them
integration tests.
