+++ Cargo.toml
@@ -248,6 +248,7 @@ include = [
     "tests/agreement_tests.rs",
     "tests/agreement_tests.txt",
     "tests/constant_time_tests.rs",
+    "tests/der_tests.rs",
     "tests/digest_tests.rs",
     "tests/digest_tests.txt",
     "tests/ecdsa_from_pkcs8_tests.txt",
